house-keeping: move /pkg to /internal
Project layout standards would name the pkg folder internal.

fixes #58
